+++ src/store/reducers/tenantOverview/executeTopTables/types.ts
@@ -1,0 +1,14 @@
+import type {IQueryResult, QueryErrorResponse} from '../../../../types/store/query';
+import type {ApiRequestAction} from '../../../utils';
+import {FETCH_TOP_TABLES, setDataWasNotLoaded} from './executeTopTables';
+
+export interface TopTablesState {
+    loading: boolean;
+    wasLoaded: boolean;
+    data?: IQueryResult;
+    error?: QueryErrorResponse;
+}
+
+export type TopTablesAction =
+    | ApiRequestAction<typeof FETCH_TOP_TABLES, IQueryResult, QueryErrorResponse>
+    | ReturnType<typeof setDataWasNotLoaded>;