+++ src/store/reducers/tenants/utils.ts
@@ -26,34 +26,44 @@ export const calculateTenantMetrics = (tenant?: TTenant) => {
         CoresUsed,
         MemoryUsed,
         StorageAllocatedSize,
-        CoresLimit,
         MemoryLimit,
-        StorageLimit,
+        StorageAllocatedLimit,
+        PoolStats,
         Metrics = {},
+        DatabaseQuotas = {},
     } = tenant || {};
 
-    const cpuFromCores = isNumeric(CoresUsed) ? Number(CoresUsed) * 1_000_000 : undefined;
-    const cpuFromMetrics = isNumeric(Metrics.CPU) ? Number(Metrics.CPU) : undefined;
+    const systemPoolUsage = PoolStats?.find(({Name}) => Name === 'System')?.Usage;
+    const userPoolUsage = PoolStats?.find(({Name}) => Name === 'User')?.Usage;
 
-    const cpu = cpuFromCores ?? cpuFromMetrics ?? undefined;
+    const cpu = isNumeric(CoresUsed) ? Number(CoresUsed) * 1_000_000 : undefined;
+    const memory = isNumeric(MemoryUsed) ? Number(MemoryUsed) : undefined;
+    const blobStorage = isNumeric(StorageAllocatedSize) ? Number(StorageAllocatedSize) : undefined;
+    const tableStorage = isNumeric(Metrics.Storage) ? Number(Metrics.Storage) : undefined;
 
-    const rawMemory = MemoryUsed ?? Metrics.Memory;
-
-    const memory = isNumeric(rawMemory) ? Number(rawMemory) : undefined;
-
-    // Blob storage - actual database size
-    const storage = isNumeric(StorageAllocatedSize) ? Number(StorageAllocatedSize) : undefined;
-    const cpuLimit = isNumeric(CoresLimit) ? Number(CoresLimit) : undefined;
+    // We use system pool usage and user pool usage to calculate cpu usage because
+    // only these pools directly indicate resources available to perform user queries
+    const cpuUsage =
+        isNumeric(systemPoolUsage) || isNumeric(userPoolUsage)
+            ? Math.max(Number(systemPoolUsage), Number(userPoolUsage)) * 100
+            : undefined;
     const memoryLimit = isNumeric(MemoryLimit) ? Number(MemoryLimit) : undefined;
-    const storageLimit = isNumeric(StorageLimit) ? Number(StorageLimit) : undefined;
+    const blobStorageLimit = isNumeric(StorageAllocatedLimit)
+        ? Number(StorageAllocatedLimit)
+        : undefined;
+    const tableStorageLimit = isNumeric(DatabaseQuotas.data_size_hard_quota)
+        ? Number(DatabaseQuotas.data_size_hard_quota)
+        : undefined;
 
     return {
         cpu,
         memory,
-        storage,
-        cpuLimit,
+        blobStorage,
+        tableStorage,
+        cpuUsage,
         memoryLimit,
-        storageLimit,
+        blobStorageLimit,
+        tableStorageLimit,
     };
 };
 
@@ -71,7 +81,7 @@ export const prepareTenants = (tenants: TTenant[], useNodeAsBackend: boolean) =>
         const backend = useNodeAsBackend ? getTenantBackend(tenant) : undefined;
         const sharedTenantName = tenants.find((item) => item.Id === tenant.ResourceId)?.Name;
         const controlPlaneName = getControlPlaneValue(tenant);
-        const {cpu, memory, storage} = calculateTenantMetrics(tenant);
+        const {cpu, memory, blobStorage} = calculateTenantMetrics(tenant);
         const {nodesCount, groupsCount} = calculateTenantEntities(tenant);
 
         return {
@@ -82,7 +92,7 @@ export const prepareTenants = (tenants: TTenant[], useNodeAsBackend: boolean) =>
             controlPlaneName,
             cpu,
             memory,
-            storage,
+            storage: blobStorage,
             nodesCount,
             groupsCount,
         };